Cambiar certificado por el del proyecto

+++ src/main/java/org/iesalixar/daw2/Alejandroangulomendez/game_zone/dtos/StudioDTO.java
@@ -19,7 +19,7 @@ public class StudioDTO {
     /**
      * Identificador único del estudio.
      */
-    private Integer id; // Ajustado a Integer para coincidir con schema.sql
+    private Long id; // Ajustado a Integer para coincidir con schema.sql
 
     /**
      * Nombre del estudio de videojuegos.
